fix: fix popal logic and rename styled-component files
