+++ toast-lib/src/Portal/index.jsx
@@ -1,20 +1,19 @@
-import { useEffect, useState } from 'react'
+import { useLayoutEffect, useRef } from 'react'
 import { createPortal } from 'react-dom'
 
 const Portal = ({ children }) => {
   const mount = document.getElementById('docs-root')
+  const el = useRef(document.createElement('div'))
 
-  const [el] = useState(() => document.createElement('div'))
-
-  useEffect(() => {
-    mount.appendChild(el)
+  useLayoutEffect(() => {
+    mount.appendChild(el.current)
 
     return () => {
-      mount.removeChild(el)
+      mount.removeChild(el.current)
     }
   }, [])
 
-  return createPortal(children, el)
+  return createPortal(children, el.current)
 }
 
 export default Portal
