+++ toast-lib/storybook-static/project.json
@@ -1,1 +1,1 @@
-{"generatedAt":1664777255913,"builder":{"name":"webpack4"},"hasCustomBabel":false,"hasCustomWebpack":false,"hasStaticDirs":false,"hasStorybookEslint":true,"refCount":0,"storybookVersion":"6.5.12","language":"javascript","storybookPackages":{"storybook":{"version":"6.5.12"},"@storybook/addon-actions":{"version":"6.5.12"},"@storybook/builder-webpack4":{"version":"6.5.12"},"@storybook/manager-webpack4":{"version":"6.5.12"},"@storybook/react":{"version":"6.5.12"},"@storybook/testing-library":{"version":"0.0.13"},"eslint-plugin-storybook":{"version":"0.6.4"}},"framework":{"name":"react"},"addons":{"@storybook/addon-links":{"version":"6.5.12"},"@storybook/addon-essentials":{"version":"6.5.12"},"@storybook/addon-interactions":{"version":"6.5.12"}}}
+{"generatedAt":1665074428006,"builder":{"name":"webpack4"},"hasCustomBabel":false,"hasCustomWebpack":true,"hasStaticDirs":false,"hasStorybookEslint":true,"refCount":0,"storybookVersion":"6.5.12","language":"javascript","storybookPackages":{"storybook":{"version":"6.5.12"},"@storybook/addon-actions":{"version":"6.5.12"},"@storybook/builder-webpack4":{"version":"6.5.12"},"@storybook/manager-webpack4":{"version":"6.5.12"},"@storybook/react":{"version":"6.5.12"},"@storybook/testing-library":{"version":"0.0.13"},"eslint-plugin-storybook":{"version":"0.6.4"}},"framework":{"name":"react"},"addons":{"@storybook/addon-links":{"version":"6.5.12"},"@storybook/addon-essentials":{"version":"6.5.12"},"@storybook/addon-interactions":{"version":"6.5.12"}}}
